+++ tech1-framework-incidents/src/test/java/io/tech1/framework/incidents/converters/impl/IncidentConverterImplTest.java
@@ -56,7 +56,8 @@ public class IncidentConverterImplTest {
         assertThat(actual.getAttributes().get("incidentType")).isEqualTo("Throwable");
         assertThat(actual.getAttributes().get("exception")).isEqualTo(NullPointerException.class);
         assertThat(actual.getAttributes().get("message")).isEqualTo("Tech1");
-        assertThat(actual.getAttributes().get("trace").toString()).startsWith("Throwable occurred! Please take required actions!");
+        assertThat(actual.getAttributes().get("trace").toString()).startsWith("java.lang.NullPointerException: Tech1");
+        assertThat(actual.getAttributes().get("trace").toString()).contains("at io.tech1.framework.incidents.converters.impl.IncidentConverterImplTest.convertThrowableIncident1Test");
     }
 
     @Test
@@ -80,7 +81,8 @@ public class IncidentConverterImplTest {
         assertThat(actual.getAttributes().get("incidentType")).isEqualTo("Throwable");
         assertThat(actual.getAttributes().get("exception")).isEqualTo(NullPointerException.class);
         assertThat(actual.getAttributes().get("message")).isEqualTo("Tech1");
-        assertThat(actual.getAttributes().get("trace").toString()).startsWith("Throwable occurred! Please take required actions!");
+        assertThat(actual.getAttributes().get("trace").toString()).startsWith("java.lang.NullPointerException: Tech1");
+        assertThat(actual.getAttributes().get("trace").toString()).contains("at io.tech1.framework.incidents.converters.impl.IncidentConverterImplTest.convertThrowableIncident2Test");
         assertThat(actual.getAttributes().get("method").toString()).contains("protected void java.lang.Object.finalize() throws java.lang.Throwable");
         assertThat(actual.getAttributes().get("params")).isEqualTo(object + ", param1, 1");
     }
@@ -105,7 +107,8 @@ public class IncidentConverterImplTest {
         assertThat(actual.getAttributes().get("incidentType")).isEqualTo("Throwable");
         assertThat(actual.getAttributes().get("exception")).isEqualTo(NullPointerException.class);
         assertThat(actual.getAttributes().get("message")).isEqualTo("Tech1");
-        assertThat(actual.getAttributes().get("trace").toString()).startsWith("Throwable occurred! Please take required actions!");
+        assertThat(actual.getAttributes().get("trace").toString()).startsWith("java.lang.NullPointerException: Tech1");
+        assertThat(actual.getAttributes().get("trace").toString()).contains("at io.tech1.framework.incidents.converters.impl.IncidentConverterImplTest.convertThrowableIncident3Test");
         assertThat(actual.getAttributes().get("key1")).isEqualTo(object);
     }